+++ .test/defaultExport/lateDefaultExportedElement.ts
@@ -1,0 +1,3 @@
+function myDefaultExportedFunction() { }
+
+export default myDefaultExportedFunction;