+++ src/models/TsDeclaration.ts
@@ -1,3 +1,4 @@
+import {TsResource} from './TsResource';
 import {CompletionItemKind} from 'vscode';
 
 export abstract class TsDeclaration {
@@ -106,7 +107,21 @@ export class ParameterDeclaration extends TsDeclaration {
     }
 }
 
-export class DefaultDeclaration extends TsDeclaration {
+export class DefaultDeclaration extends TsExportableDeclaration {
+    private exported: TsDeclaration;
+
+    public get exportedDeclaration(): TsDeclaration {
+        if (!this.exported) {
+            this.exported = this.resource.declarations.find(o => o.name === this.name);
+        }
+
+        return this.exported;
+    }
+
+    constructor(name: string, private resource: TsResource) {
+        super(name, true);
+    }
+
     public getItemKind(): CompletionItemKind {
         return CompletionItemKind.File;
     }