+++ src/utilities/ResolveIndexExtensions.ts
@@ -8,20 +8,30 @@ export function getDeclarationsFilteredByImports(resolveIndex: ResolveIndex, doc
     let declarations = resolveIndex.declarationInfos;
 
     for (let tsImport of imports) {
+            let importedLib = getAbsolutLibraryName(tsImport.libraryName, documentPath);
         if (tsImport instanceof TsNamedImport) {
-            let importedLib = getRelativeLibraryName(tsImport.libraryName, documentPath);
             declarations = declarations.filter(o => o.from !== importedLib || !tsImport.specifiers.some(s => s.specifier === o.declaration.name));
         } else if (tsImport instanceof TsNamespaceImport || tsImport instanceof TsExternalModuleImport) {
             declarations = declarations.filter(o => o.from !== tsImport.libraryName);
         } else if (tsImport instanceof TsDefaultImport) {
-            declarations = declarations.filter(o => (!(o.declaration instanceof DefaultDeclaration) || tsImport.libraryName !== o.from));
+            declarations = declarations.filter(o => (!(o.declaration instanceof DefaultDeclaration) || importedLib !== o.from));
         }
     }
 
     return declarations;
 }
 
-export function getRelativeLibraryName(library: string, actualFilePath: string): string {
+/**
+ * Returns the absolut workspace specific library path.
+ * If the library is a node module or a typings module, the name
+ * is returned. If the "lib" is in the local workspace, then the
+ * absolut path from the workspace root is returned.
+ * 
+ * @param {string} library - Name of the library
+ * @param {string} actualFilePath - Filepath of the actually open file
+ * @returns {string} - Absolut path from the workspace root to the desired library
+ */
+export function getAbsolutLibraryName(library: string, actualFilePath: string): string {
     if (!library.startsWith('.')) {
         return library;
     }
@@ -29,7 +39,17 @@ export function getRelativeLibraryName(library: string, actualFilePath: string):
     return relative;
 }
 
-export function getRelativeImportPath(library: string, actualFilePath: string): string {
+/**
+ * Returns the relative path to a specific library.
+ * If the library is a node module or a typings module, the name
+ * is returned. If the "lib" is in the local workspace, then the
+ * relative path from the actual file is returned.
+ * 
+ * @param {string} library - Name of the library
+ * @param {string} actualFilePath - Filepath of the actually open file
+ * @returns {string} - Relative path from the actual file to the library
+ */
+export function getRelativeLibraryName(library: string, actualFilePath: string): string {
     if (!library.startsWith('/')) {
         return library;
     }